Made missle shot slightly more interesting

+++ Assets/Scripts/player/PlayerMissile.cs
@@ -7,7 +7,7 @@ public class PlayerMissile : PlayerProjectile
     
     [SerializeField] private float xBoundary = 30;
     [SerializeField] private float playerXOffset = 0.5f;
-    [SerializeField] private int playerYOffset = 2;
+    [SerializeField] private float playerYOffset = 2;
     [SerializeField] private float playerZOffset = -0.2f;
 
     protected override void Start()
